+++ CotEditor/Sources/DefaultKeys.swift
@@ -181,6 +181,7 @@ extension DefaultKeys {
     static let savesTextOrientation = DefaultKey<Bool>("savesTextOrientation")
     static let maximumRecentStyleCount = DefaultKey<Int>("maximumRecentStyleCount")
     static let maximumSelectionInstanceHighlightCount = DefaultKey<Int>("maximumSelectionInstanceHighlightCount")
+    static let minimumLengthForNonContiguousLayout = DefaultKey<Int>("minimumLengthForNonContiguousLayout")
     
     static let lastVersion = DefaultKey<String?>("lastVersion")