+++ CotEditor/Sources/DefaultSettings.swift
@@ -186,6 +186,7 @@ struct DefaultSettings {
         .savesTextOrientation: true,
         .maximumRecentStyleCount: 6,
         .maximumSelectionInstanceHighlightCount: 100,
+        .minimumLengthForNonContiguousLayout: 5_000_000,
     ]