+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Change Log
 - Add steppers to the font setting controls.
 - Restore the characters even incompatible with the document encoding when restoring documents from the last session.
 - Optimize syntax parsing.
+- Switch the text layout mode to non-contiguous only when the document is large.
 - Prefer using .yml for syntax definition files over .yaml.
 - Deprecate the feature to replace `$LN` in the outline menu template with the line number of the occurrence.
 - Remove original document icons for CoffeeScript and Tcl.